app/requirements.txt
streamlit
google-cloud-bigquery
google-cloud-storage